refactor: merge extractSynonym and extractAntonym
